move findOrBuildResourceEntity tests from /authorization to /utils
